+++ src/test/kotlin/com/forgerock/uk/openbanking/tests/functional/account/statements/api/v3_1_8/GetStatements.kt
@@ -1,0 +1,35 @@
+package com.forgerock.uk.openbanking.tests.functional.account.statements.api.v3_1_8
+
+import assertk.assertThat
+import assertk.assertions.isNotEmpty
+import assertk.assertions.isNotNull
+import com.forgerock.securebanking.framework.extensions.junit.CreateTppCallback
+import com.forgerock.securebanking.openbanking.uk.common.api.meta.obie.OBVersion
+import com.forgerock.uk.openbanking.support.account.AccountRS
+import com.forgerock.uk.openbanking.tests.functional.account.access.BaseAccountApi3_1_8
+import uk.org.openbanking.datamodel.account.OBExternalPermissions1Code
+import uk.org.openbanking.datamodel.account.OBReadStatement2
+
+class GetStatements(version: OBVersion, tppResource: CreateTppCallback.TppResource) :
+    BaseAccountApi3_1_8(version, tppResource) {
+
+    fun shouldGetStatementsTest() {
+        // Given
+        val permissions = listOf(
+            OBExternalPermissions1Code.READSTATEMENTSBASIC,
+            OBExternalPermissions1Code.READSTATEMENTSDETAIL,
+            OBExternalPermissions1Code.READACCOUNTSDETAIL
+        )
+        val (_, accessToken) = accountAccessConsentApi.createConsentAndGetAccessToken(permissions)
+
+        // When
+        val result = AccountRS().getAccountsData<OBReadStatement2>(
+            accountsApiLinks.GetStatements,
+            accessToken
+        )
+
+        // Then
+        assertThat(result).isNotNull()
+        assertThat(result.data.statement).isNotEmpty()
+    }
+}
