+++ src/test/kotlin/com/forgerock/uk/openbanking/tests/functional/account/statements/legacy/LegacyGetAccountStatementTest.kt
@@ -1,4 +1,4 @@
-package com.forgerock.uk.openbanking.tests.functional.account.statements
+package com.forgerock.uk.openbanking.tests.functional.account.statements.legacy
 
 import assertk.assertThat
 import assertk.assertions.isNotEmpty
@@ -11,14 +11,13 @@ import com.forgerock.uk.openbanking.support.account.AccountFactory.Companion.obR
 import com.forgerock.uk.openbanking.support.account.AccountRS
 import com.forgerock.uk.openbanking.support.discovery.accountAndTransaction3_1
 import com.forgerock.uk.openbanking.support.discovery.accountAndTransaction3_1_2
-import com.forgerock.uk.openbanking.support.discovery.accountAndTransaction3_1_8
 import org.junit.jupiter.api.Test
 import uk.org.openbanking.datamodel.account.OBExternalPermissions1Code
 import uk.org.openbanking.datamodel.account.OBReadAccount3
 import uk.org.openbanking.datamodel.account.OBReadConsentResponse1
 import uk.org.openbanking.datamodel.account.OBReadStatement2
 
-class GetAccountStatementTest(val tppResource: CreateTppCallback.TppResource) {
+class LegacyGetAccountStatementTest(val tppResource: CreateTppCallback.TppResource) {
 
     @EnabledIfVersion(
         type = "accounts",
@@ -66,67 +65,6 @@ class GetAccountStatementTest(val tppResource: CreateTppCallback.TppResource) {
         assertThat(resultGetAccountStatements.data.statement[0].statementId).isNotEmpty()
 
         val accountDataUrl = accountAndTransaction3_1_2.Links.links.GetAccountStatement
-            .replace("{AccountId}", accounts.data.account[0].accountId)
-            .replace("{StatementId}", resultGetAccountStatements.data.statement[0].statementId)
-
-        // When
-        val result = AccountRS().getAccountData<OBReadStatement2>(
-            accountDataUrl,
-            accessToken,
-            accounts.data.account[0].accountId
-        )
-
-        // Then
-        assertThat(result).isNotNull()
-        assertThat(result.data.statement).isNotEmpty()
-    }
-
-    @EnabledIfVersion(
-        type = "accounts",
-        apiVersion = "v3.1.8",
-        operations = ["CreateAccountAccessConsent", "GetAccounts", "GetAccountStatement"],
-        apis = ["statements"],
-        compatibleVersions = ["v.3.1.7", "v.3.1.6", "v.3.1.5", "v.3.1.4", "v.3.1.3"]
-    )
-    @Test
-    fun shouldGetAccountStatement_v3_1_8() {
-        // Given
-        val consentRequest = obReadConsent1(
-            listOf(
-                OBExternalPermissions1Code.READSTATEMENTSBASIC,
-                OBExternalPermissions1Code.READSTATEMENTSDETAIL,
-                OBExternalPermissions1Code.READACCOUNTSDETAIL
-            )
-        )
-        val consent = AccountRS().consent<OBReadConsentResponse1>(
-            accountAndTransaction3_1_8.Links.links.CreateAccountAccessConsent,
-            consentRequest,
-            tppResource.tpp
-        )
-        val accessToken = AccountAS().getAccessToken(
-            consent.data.consentId,
-            tppResource.tpp.registrationResponse,
-            psu,
-            tppResource.tpp
-        )
-        val accounts =
-            AccountRS().getAccountsData<OBReadAccount3>(accountAndTransaction3_1_8.Links.links.GetAccounts, accessToken)
-
-        assertThat(accounts.data).isNotNull()
-        assertThat(accounts.data.account[0].accountId).isNotNull()
-
-        val accountStatementDataUrl = accountAndTransaction3_1_8.Links.links.GetAccountStatements
-            .replace("{AccountId}", accounts.data.account[0].accountId)
-        val resultGetAccountStatements = AccountRS().getAccountData<OBReadStatement2>(
-            accountStatementDataUrl,
-            accessToken,
-            accounts.data.account[0].accountId
-        )
-
-        assertThat(resultGetAccountStatements).isNotNull()
-        assertThat(resultGetAccountStatements.data.statement[0].statementId).isNotEmpty()
-
-        val accountDataUrl = accountAndTransaction3_1_8.Links.links.GetAccountStatement
             .replace("{AccountId}", accounts.data.account[0].accountId)
             .replace("{StatementId}", resultGetAccountStatements.data.statement[0].statementId)
