+++ src/test/kotlin/com/forgerock/uk/openbanking/tests/functional/account/statements/legacy/LegacyGetStatementsTest.kt
@@ -1,4 +1,4 @@
-package com.forgerock.uk.openbanking.tests.functional.account.statements
+package com.forgerock.uk.openbanking.tests.functional.account.statements.legacy
 
 import assertk.assertThat
 import assertk.assertions.isNotEmpty
@@ -10,13 +10,12 @@ import com.forgerock.uk.openbanking.support.account.AccountAS
 import com.forgerock.uk.openbanking.support.account.AccountFactory.Companion.obReadConsent1
 import com.forgerock.uk.openbanking.support.account.AccountRS
 import com.forgerock.uk.openbanking.support.discovery.accountAndTransaction3_1_2
-import com.forgerock.uk.openbanking.support.discovery.accountAndTransaction3_1_8
 import org.junit.jupiter.api.Test
 import uk.org.openbanking.datamodel.account.OBExternalPermissions1Code
 import uk.org.openbanking.datamodel.account.OBReadConsentResponse1
 import uk.org.openbanking.datamodel.account.OBReadStatement2
 
-class GetStatementsTest(val tppResource: CreateTppCallback.TppResource) {
+class LegacyGetStatementsTest(val tppResource: CreateTppCallback.TppResource) {
 
     @EnabledIfVersion(
         type = "accounts",
@@ -50,46 +49,6 @@ class GetStatementsTest(val tppResource: CreateTppCallback.TppResource) {
         // When
         val result = AccountRS().getAccountsData<OBReadStatement2>(
             accountAndTransaction3_1_2.Links.links.GetStatements,
-            accessToken
-        )
-
-        // Then
-        assertThat(result).isNotNull()
-        assertThat(result.data.statement).isNotEmpty()
-    }
-
-    @EnabledIfVersion(
-        type = "accounts",
-        apiVersion = "v3.1.8",
-        operations = ["CreateAccountAccessConsent", "GetAccounts", "GetStatements"],
-        apis = ["statements"],
-        compatibleVersions = ["v.3.1.7", "v.3.1.6", "v.3.1.5", "v.3.1.4", "v.3.1.3"]
-    )
-    @Test
-    fun shouldGetStatements_v3_1_8() {
-        // Given
-        val consentRequest = obReadConsent1(
-            listOf(
-                OBExternalPermissions1Code.READSTATEMENTSBASIC,
-                OBExternalPermissions1Code.READSTATEMENTSDETAIL,
-                OBExternalPermissions1Code.READACCOUNTSDETAIL
-            )
-        )
-        val consent = AccountRS().consent<OBReadConsentResponse1>(
-            accountAndTransaction3_1_8.Links.links.CreateAccountAccessConsent,
-            consentRequest,
-            tppResource.tpp
-        )
-        val accessToken = AccountAS().getAccessToken(
-            consent.data.consentId,
-            tppResource.tpp.registrationResponse,
-            psu,
-            tppResource.tpp
-        )
-
-        // When
-        val result = AccountRS().getAccountsData<OBReadStatement2>(
-            accountAndTransaction3_1_8.Links.links.GetStatements,
             accessToken
         )
